+++ projects/ngx-translation-validation/src/lib/ngx-tv.errors.ts
@@ -1,0 +1,6 @@
+export class ControlNameNotFoundError extends Error {
+  constructor(reason: string) {
+    super(`Name for control cannot be found: ${reason}`);
+    this.name = 'ControlNameNotFoundError';
+  }
+}