+++ projects/ngx-translation-validation/src/lib/test/transloco-testing.module.ts
@@ -1,0 +1,13 @@
+import { TranslocoTestingModule, TranslocoTestingOptions } from '@ngneat/transloco';
+
+export function getTranslocoTestingModule(options: TranslocoTestingOptions = {}) {
+  return TranslocoTestingModule.forRoot({
+    langs: { tk: {} },
+    translocoConfig: {
+      availableLangs: ['tk'],
+      defaultLang: 'tk',
+    },
+    preloadLangs: true,
+    ...options,
+  });
+}