+++ projects/ngx-translation-validation/src/public-api.ts
@@ -10,3 +10,4 @@ export * from './lib/ngx-tv.config';
 export * from './lib/ngx-tv-container/ngx-tv-container.component';
 export * from './lib/ngx-tv.module';
 export * from './lib/ngx-tv.providers';
+export * from './lib/ngx-tv.errors';