+++ src/app/app.component.ts
@@ -2,13 +2,14 @@ import { Component, inject } from '@angular/core';
 import { FormsModule, NonNullableFormBuilder, ReactiveFormsModule, Validators } from '@angular/forms';
 import { TranslocoDirective, TranslocoService } from '@ngneat/transloco';
 import { NgxTvModule } from 'ngx-translation-validation';
+import { NgFor } from '@angular/common';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss'],
   standalone: true,
-  imports: [TranslocoDirective, FormsModule, NgxTvModule, ReactiveFormsModule],
+  imports: [TranslocoDirective, FormsModule, NgxTvModule, ReactiveFormsModule, NgFor],
 })
 export class AppComponent {
   private readonly formBuilder = inject(NonNullableFormBuilder);
@@ -46,6 +47,17 @@ export class AppComponent {
       updateOn: 'change',
     },
   );
+
+  protected readonly formArray = this.formBuilder.group({
+    persons: this.formBuilder.array([
+      this.formBuilder.group({
+        name: ['', Validators.required],
+        email: ['', Validators.email],
+        phone: ['', [Validators.required, Validators.maxLength(12)]],
+      }),
+    ]),
+    checks: this.formBuilder.array([[false, Validators.requiredTrue], [true, Validators.requiredTrue], [false]]),
+  });
 
   protected onSubmit(): void {
     // Intentionally left blank to trigger validation